Fix p2p reshape synchronize after dfe2e61883ee6db8a8a1dff0733c7e67c586a415
Turns out synchrnization is actually required for p2p reshapes, because
the next reshape could overwrite the scratch space of a running
kernel without it. This probably should be done in a more fine grained manner,
because synchronizing after an fft should not be required as all their
operations follow stream ordering.